requirements.txt
flask
langchain
openai
python-dotenv
faiss-cpu
pyyaml
sqlite3